Put prepareData into a module
