+++ src/index.js
@@ -1,36 +1,7 @@
 import parseOptions from './options';
+import prepareData from './data';
 import { prepareTemplates } from './template';
-import * as utils from './utils';
 
-/**
- * Build a data/context object for use by the builder
- * @TODO This may move into its own module if it seems appropriate
- * @TODO Make this testable
- *
- * @param {Object} options
- * @return {Promise} resolving to {Object} of keyed file data
- */
-function prepareData (options) {
-  var parseFiles = utils.readFilesKeyed;
-  // Data data
-  const dataData = parseFiles(options.data, {
-    contentFn: options.data.parseFn
-  });
-  // Layouts data
-  const layoutData = parseFiles(options.layouts);
-
-  // Page data
-  // parseFiles (options.pages) with a content fn of
-  // front matter
-
-  // Build data object
-  return Promise.all([dataData, layoutData])
-    .then(allData => {
-      return { data: allData[0],
-               layouts: allData[1]
-             };
-    });
-}
 /**
  * Build the drizzle output
  *
